requirements.txt
python-dotenv==1.0.1
requests==2.32.3
translate==3.6.1
google-generativeai==0.8.5
googlesearch-python==1.3.0
